requirements.txt
tensorflow>=2.0
numpy
pyyaml
pooltool